fix: Nix compatibility
Nix is a reproducible, deterministic, declarative package manager.
Variable compile time inputs/outputs defeat the purpose.  This wraps the
git revision and datetime inputs in a check for an environment variable
present in the Nix build environment.  If so, the Nix package manager
will provide the git revision correctly within its environment and a
static datetime equal to the start of the Unix epoch (1970-01-01
00:00:00).